pyproject.toml
[project]
name = "newsletter-curator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.79.0",
    "apscheduler>=3.10",
    "azure-identity>=1.25.1",
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "msgraph-sdk>=1.54.0",
    "notion-client>=2.7.0",
    "playwright>=1.40.0",
    "python-dotenv>=1.2.1",
    "rapidfuzz>=3.0.0",
    "reflex>=0.8.26",
    "trafilatura>=2.0.0",
]
